rename childAccountPolicy folder to childAccount
